Fixes #439 inject uploadcare
