+++ example/index.html
@@ -4,8 +4,6 @@
     <meta charset="utf-8">
     <title>Example Scene</title>
     <script src="js/aframe-master.js"></script>
-    <script>UPLOADCARE_PUBLIC_KEY = "f43ad452b58f9e853d05";</script>
-    <script src="https://ucarecdn.com/widget/2.10.0/uploadcare/uploadcare.full.min.js" charset="utf-8"></script>
   </head>
   <body>
     <a-scene debug="true">
