+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aframe-inspector",
-  "version": "0.5.0",
+  "version": "0.5.1",
   "description": "A visual inspector tool for A-Frame.",
   "main": "dist/aframe-inspector.min.js",
   "scripts": {
